Initial commit for the Address book
This program is an outline of what is to be developed. The
add_entries function has been partially developed. It still needs
to be improved to check for data intergrity, and to also search the
database (that is the Contact.txt file), to check if the entry exist
and enable editing if it exist.
Search and Editing will be implemented, after the search and edit
functions have been implemented. Because, they will be called by the
add_entries functions to perform Search and Editing respectively.