Implement the search function
The changes made are only on the 'search' function. The 'add_entries'
function that was implemented earlier, is unchanged.
the 'search' function can now search for lines, that have the given
pattern and stores the results in a variable called 'results'.
Improvements to be made to the 'search' function include:
    - Restricting search to names only,
    - Enabling the user to select one search result and,
    - How to respond if nothing is found.