Make it possible to select one search result
The 'search' function has been improved to make it 'echo' all
search results, with a unique 'index' next to it. The 'index' is then
used to select one of the results. If only one result is found it is
automatically displayed.
If only one result is found, It is automatically displayed.
If no result is found, a statement to that effect is 'echoed.'