Link the add_entries, search and edit functions
The add_entries function has been improved to check on data
intergrity of the names, phone numbers and emails input by the user.
It also calls the search function to confirm that their isn't any
existing contact with the same name.
The search has been improved to restrict search to the name only,
and also make it easier, for it to be called by the edit and add_entries
functions. It also has customized responses, depending on whether it was
called directly or by another function.
The edit function has also been improved to allow it to be called by
other functions. It also calls the search function every time it needs
to search for something.
Lastly, It allows the user to input the new name, replacing the existing
one.