Add main function that controls the program calls
A main fuction has been implemented. It accepts commands from the user and
calls the appropriate function to execute them. It also provides the possible
commands, if an illegal command is entered, and the help command provides
information on the commands.
The 'edit' and 'delete' functions have been improved, to make them check, if
a name exists, and asking the user to confirm, before editing or deleting
respectively.